+++ plasma_framework/contracts/src/exits/payment/controllers/PaymentChallengeIFEInputSpent.sol
@@ -77,7 +77,7 @@ library PaymentChallengeIFEInputSpent {
     {
         require(args.senderData == keccak256(abi.encodePacked(msg.sender)), "Incorrect senderData");
 
-        uint160 exitId = ExitId.getInFlightExitId(args.inFlightTx);
+        uint168 exitId = ExitId.getInFlightExitId(args.inFlightTx);
         PaymentExitDataModel.InFlightExit storage ife = inFlightExitMap.exits[exitId];
 
         require(ife.exitStartTimestamp != 0, "In-flight exit does not exist");
